+++ lib/components/ConditionalWrapper/ConditionalWrapper.d.ts
@@ -1,0 +1,5 @@
+export declare const ConditionalWrapper: ({ condition, wrapper, children }: {
+    condition: any;
+    wrapper: any;
+    children: any;
+}) => any;