+++ src/lib/components/index.ts
@@ -3,6 +3,7 @@ export * from './Backdrop';
 export * from './Button';
 export * from './Breadcrumb';
 export * from './Card';
+export * from './ConditionalWrapper';
 export * from './Checkbox';
 export * from './Drawer';
 export * from './DropDown';